requirements.txt
alembic==1.9.2
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.5
cachelib==0.10.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
decorator==5.1.1
filelock==3.9.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.3
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gevent==22.10.2
greenlet==2.0.1
idna==3.4
infinity==1.5
install==1.3.5
intervals==0.9.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
mypy==0.991
mypy-extensions==0.4.3
Pillow==9.4.0
prompt-toolkit==3.0.37
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
redis==4.5.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.1
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
utils==1.0.1
validators==0.20.0
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
WTForms-Components==0.10.5
zope.event==4.6
zope.interface==5.5.2
